move cvr handling to cvr handler
